Moves reducing_adapter class to ygm::container::detail
