+++ include/ygm/container/detail/reducing_adapter.hpp
@@ -8,7 +8,7 @@
 #include <ygm/detail/ygm_ptr.hpp>
 #include <ygm/detail/ygm_traits.hpp>
 
-namespace ygm::container {
+namespace ygm::container::detail {
 
 template <typename Container, typename ReductionOp>
 class reducing_adapter {
@@ -118,4 +118,4 @@ reducing_adapter<Container, ReductionOp> make_reducing_adapter(
   return reducing_adapter<Container, ReductionOp>(c, reducer);
 }
 
-}  // namespace ygm::container
+}  // namespace ygm::container::detail
